refactor: remove workspace related migration code
