+++ crates/nebula-domain/src/database/authorization_migration/mod.rs
@@ -1,31 +1,33 @@
 use async_trait::async_trait;
-use sea_orm::DbErr;
+use sea_orm::{
+    sea_query::{Alias, IntoIden},
+    DbErr,
+};
 use sea_orm_migration::{IntoSchemaManagerConnection, MigrationTrait, MigratorTrait};
 
 use super::AuthMethod;
 
-mod m20241126_001_init_backbone;
+mod m20241128_001_init_authorization;
 
 pub struct Migrator;
 
 #[async_trait]
 impl MigratorTrait for Migrator {
     fn migrations() -> Vec<Box<dyn MigrationTrait>> {
-        vec![Box::new(m20241126_001_init_backbone::Migration)]
+        vec![Box::new(m20241128_001_init_authorization::Migration)]
+    }
+
+    fn migration_table_name() -> sea_orm::DynIden {
+        Alias::new("seaql_migrations_authorization").into_iden()
     }
 }
 
-#[cfg(test)]
-pub async fn migrate_backbone(_host: &str, _port: u16, _database_name: &str, _auth: &AuthMethod) -> anyhow::Result<()> {
-    use tracing::debug;
-
-    debug!("workspace migration not supported in test environment");
-
-    Ok(())
-}
-
-#[cfg(not(test))]
-pub async fn migrate_backbone(host: &str, port: u16, database_name: &str, auth: &AuthMethod) -> anyhow::Result<()> {
+pub async fn migrate_authorization(
+    host: &str,
+    port: u16,
+    database_name: &str,
+    auth: &AuthMethod,
+) -> anyhow::Result<()> {
     let connection = super::connect_to_database(host, port, database_name, auth).await?;
 
     migrate(connection.as_ref()).await?;
