--- crates/nebula-domain/src/database/workspace.rs
@@ -1,19 +1,0 @@
-use chrono::{DateTime, Utc};
-use sea_orm::prelude::*;
-
-use super::UlidId;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
-#[sea_orm(table_name = "workspace")]
-pub struct Model {
-    #[sea_orm(primary_key)]
-    pub id: UlidId,
-    pub name: String,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {}
-
-impl ActiveModelBehavior for ActiveModel {}
